+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Added a config option to enable TE Machine Frames in the recipes
 + Added a blacklist for the mob duplicator, blacklisted entities are shown when captured in the Mob Imprisonment Tools
 + Changed a color in the StoneWork Factory so it's not shared with the power one
++ Wither Builders use items from any slot if they need it
 
 # 1.5.12
 + Fixed BHU being derpy with the new ItemHandler for big stacks